testing_requirements.txt
argparse==1.2.1
blinker==1.2
coverage==3.6
Flask-Testing==0.4
nose==1.3.0
wsgiref==0.1.2
